+++ secrets/ed25519_test.go
@@ -1,0 +1,7 @@
+package secrets
+
+import "testing"
+
+func TestEd25519(t *testing.T) {
+	// TODO: GenerateEd25519Keys
+}