+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.3.33 - Ed25519 Signature 
+  * Commit: 
+  * Generate Ed25519 Keys
 ## v0.3.32 - Combinations
   * Commit: 2026-05-22 06:59 
   * Combinations